Add basic enum for PieceTypes
There are 20 unique pieces types in the Gungi board game. A piece consists of a black and white chip. On one side is a "Front" type piece and on the back are the "Back" type pieces. Every Front Type piece (except for Commander) has a Back Type piece.

Gungi Rules: https://mmmmalo.tumblr.com/post/74510568781/rules-of-gungi