Don't panic on bad inputs for set, drop and pop
Towers no longer panic when trying to use one of these functions incorrectly. They return a Result instead. Set also has now been allowed to make the tower state invalid (this is because I realized that it might be cumbersome to force the tower to remain in a good state constantly).