Convert Tower struct into a Tower enum
Using three fields like a vector makes the struct very hard to use. Additionally, having all three fields be Option meant doing a lot of unwraps all over the place, which is annoying. Using a vector or array to store the three pieces also would result in the same issue. We fix this by unifying the Tower height with the Tower struct to produce four enums, one for each possible height of the Tower (Empty, Bottom, Middle, and Top).