Add `get_move_special`
`get_move_special` takes a PieceType, Color, and Tier and returns the appropriate MoveSpecial associated with them.

This commit adds the concept of a "Tier" as well. Pieces can be Tiers 1, 2, or 3. The Tier they are is determined by what position they are in a Tower. Tiers can change through out the game and often do. Pieces of different Tiers may move in different ways.